MODSOURMAN-1063: Test classes refactoring.
